fixes player interacting whilst in conversation
