+++ Assets/Scripts/NPCs/DialogueNPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ee6938cefe6a5e45bfbc59ca2a426a6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
